requirements.txt
boto3
#jinja2
ruamel.yaml
toml
yq
